dependency updated. Whent for an install requires in setup.cfg
Removed requirements.txt to avoid defining deps in two places.
