--- requirements.txt
@@ -1,9 +1,0 @@
-certifi==2025.7.14
-charset-normalizer==3.4.2
-idna==3.10
-packaging==25.0
-PyQt6==6.9.1
-PyQt6-Qt6==6.9.1
-PyQt6_sip==13.10.2
-requests==2.32.4
-urllib3==2.5.0
