+++ setup.cfg
@@ -29,15 +29,11 @@ packages = find:
 include_package_data = True
 python_requires = >=3.8
 install_requires =
-    certifi==2025.7.14
-    charset-normalizer==3.4.2
-    idna==3.10
-    packaging==25.0
     PyQt6==6.9.1
-    PyQt6-Qt6==6.9.1
-    PyQt6_sip==13.10.2
     requests==2.32.4
-    urllib3==2.5.0
+    httpx==0.28.1
+    packaging==25.0
+
 
 [options.packages.find]
 where=src
@@ -53,7 +49,7 @@ dev =
 
 [options.entry_points]
 gui_scripts =
-    gambit-pairing = gambitpairing.app:main
+    gambit-pairing = gambitpairing:main
 
 [project_urls]
 Homepage = https://www.chickaboo.net/gambit-pairing/
