+++ src/resources/styles.qss
